+++ app/controllers/api/v1/user_settings_controller.rb
@@ -17,7 +17,7 @@ module Api
       api :PATCH, '/v1/user_settings/update_settings.json', 'Update user settings'
       error code: 401, desc: 'Unauthorized'
       def update_settings
-        UpdateUserTimeOffset.call(user: Current.user, value: user_settings_params[:time_offset][:value])
+        UpdateTimeOffset.call(timeable: Current.user, value: user_settings_params[:time_offset][:value])
         render json: { result: 'User settings are updated' }, status: 200
       end