+++ app/form_objects/time_offset_form.rb
@@ -4,11 +4,14 @@ class TimeOffsetForm
   include Virtus.model
 
   attribute :id, Integer
-  attribute :user, User
+  attribute :timeable_id, Integer
+  attribute :timeable_type, String
   attribute :value, Integer, default: nil
 
-  validates :user, presence: true
+  validates :timeable_id, :timeable_type, presence: true
+  validates :timeable_type, inclusion: { in: %w[User Guild] }
   validates :value, inclusion: -12..12, allow_nil: true
+  validate :timeable_exists?
 
   attr_reader :time_offset
 
@@ -19,5 +22,13 @@ class TimeOffsetForm
     @time_offset.attributes = attributes.except(:id)
     @time_offset.save
     true
+  end
+
+  private
+
+  def timeable_exists?
+    return unless timeable_type.present?
+    return if timeable_type.constantize.where(id: timeable_id).exists?
+    errors[:timeable] << 'is not exists'
   end
 end