+++ app/models/concerns/timeable.rb
@@ -1,0 +1,8 @@
+# Represents deliveriable
+module Timeable
+  extend ActiveSupport::Concern
+
+  included do
+    has_one :time_offset, as: :timeable, dependent: :destroy
+  end
+end