+++ app/serializers/guild_show_serializer.rb
@@ -1,3 +1,7 @@
 class GuildShowSerializer < ActiveModel::Serializer
-  attributes :id, :name, :description, :slug, :locale
+  attributes :id, :name, :description, :slug, :locale, :time_offset_value
+
+  def time_offset_value
+    object.time_offset&.value
+  end
 end