+++ spec/factories/time_offsets.rb
@@ -1,6 +1,6 @@
 FactoryBot.define do
   factory :time_offset do
     value { nil }
-    association :user
+    association :timeable, factory: :user
   end
 end