+++ spec/interactors/update_time_offset_spec.rb
@@ -1,9 +1,9 @@
-describe UpdateUserTimeOffset do
+describe UpdateTimeOffset do
   let!(:user) { create :user }
 
   describe '.call' do
     context 'for empty value' do
-      let(:interactor) { UpdateUserTimeOffset.call(user: user, value: '') }
+      let(:interactor) { described_class.call(timeable: user, value: '') }
 
       it 'succeeds' do
         expect(interactor).to be_a_success
@@ -11,14 +11,14 @@ describe UpdateUserTimeOffset do
 
       it 'update value param to nil' do
         interactor
-        user.time_offset.reload
+        user.reload
 
         expect(user.time_offset.value).to eq nil
       end
     end
 
     context 'for not empty value' do
-      let(:interactor) { UpdateUserTimeOffset.call(user: user, value: '3') }
+      let(:interactor) { described_class.call(timeable: user, value: '3') }
 
       it 'succeeds' do
         expect(interactor).to be_a_success
@@ -26,7 +26,7 @@ describe UpdateUserTimeOffset do
 
       it 'update value param to 3' do
         interactor
-        user.time_offset.reload
+        user.reload
 
         expect(user.time_offset.value).to eq 3
       end