+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - locale to guilds
+- time_offset to guilds
 
 ## [1.1.4] - 2019-12-18
 ### Added